requirements.txt
openai
ollama
python-dotenv
rich
pandas
chromadb>=0.4.22
groq>=0.4.0